+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- If CAPMC `status` function is called with an empty node list, log a warning and return without
+  calling CAPMC.
 
 ## [2.0.32] - 03-21-2024
 ### Fixed